tss package: refactor+added symkey creating via DH
